requirements.txt
boto3==1.9.71
botocore==1.12.71
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-storages==1.7.1
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
docutils==0.14
firebase-admin==2.14.0
Flask==1.0.2
gcloud==0.17.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jws==0.1.3
Markdown==3.0.1
MarkupSafe==1.1.0
msgpack==0.6.0
mysqlclient==1.3.14
oauth2client==3.0.0
Pillow==5.3.0
protobuf==3.6.1
psycopg2==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
PyJWT==1.7.1
python-dateutil==2.7.5
python-jwt==2.0.1
pytz==2018.7
requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
whitenoise==4.1.2
